- Fixed datatype issues in the MobInfo struct
git-svn-id: svn://70.113.44.214/VanaMirror/trunk@1345 09c1253f-3d09-8e40-9719-4320e56601b5

+++ src/ChannelServer/Mobs.h
@@ -44,10 +44,10 @@ struct MobAttackInfo {
 };
 
 struct MobInfo {
-	uint16_t hp;
-	uint16_t mp;
-	uint16_t hprecovery;
-	uint16_t mprecovery;
+	uint32_t hp;
+	uint32_t mp;
+	uint32_t hprecovery;
+	uint32_t mprecovery;
 	int32_t exp;
 	int8_t hpcolor;
 	int8_t hpbgcolor;
